[Agent] Refactor LLM config loader
